spi: Require unassigned pins for all functions
On a closer read of the switch matrix documentation, all of the SPI pins
are considered bidirectional regardless of the current mode of the SPI
periperhals. In retrospect that makes sense because otherwise the switch
matrix would need to know what mode the SPI peripherals are in and might
end up needing to tie-break a bunch of selections if the mode changed.

Therefore we only need one implementation of each of the pin-selection
methods now, because they can apply to any active mode.